Update support for being built in rust-lang/rust
This commit updates the support necessary for this crate to be built as
a dependency of `getopts` which is a dependency of the `test` crate
which is built in rust-lang/rust. This change will be required to land
rust-lang/rust#63637 where Rust's build system is being tweaked
slightly. This is intended to not have any impact on other users of this
crate and the `Cargo.toml` features here should only be used by Rust's
own build system.